requirements.txt
Django==4.1.1
djangorestframework==3.14.0
pydantic==1.10.2
PyJWT==2.6.0
tornado==6.2
twisted==22.10.0
corsheaders==3.13.0
djangorestframework-simplejwt==5.2.2
